requirements.txt
openai
pillow
PyQt6
pywin32
wechat_ocr
keyboard